usb: ci13xxx_udc: Add missing ifdef/endif
When CONFIG_USB_GADGET_DEBUG_FILES is disabled, the label del_udc still
gets compiled, hence making the compiler complain that del_udc is defined
but not used ([-Wunused-label]).

To fix that, the code must be wrapped around with #ifdef CONFIG... #endif
as the other code for CONFIG_USB_GADGET_DEBUG_FILES in the same function.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>