drivers: Clean import rosy touchscreen drivers
- Originally based on: LA.UM.6.6.r1-09900-89xx.0

Debugging is reduced to its bare minimum.
The configuration of the drivers is optimized.
FT8006M is now using its own pramboot file, as it should be.
Pramboot version of FT8716 is updated to the latest one found.

The touchscreen freezing issue that users randomly experience
is now gone. The cause for it was the increased debugging that,
by default, should not be enabled in production.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>